Folders

+++ Primitives/Networking/Connscan/Connscan.py
@@ -14,7 +14,8 @@ count = 0
 index = 0
 arrayOfFiles=['Offset.txt','Local-Address.txt','Remote-Address.txt','Pid.txt','temp.txt']
 for x in arrayOfFiles:
-	os.remove(x)
+	if os.path.exists(x):
+		os.remove(x)
 for oneLine in output.split("\n"):
 	if count < 2:
 		count = count+1
