+++ Primitives/Registry/Hivelist/HiveList.py
@@ -15,7 +15,8 @@ index = 0
 arrayOfFiles=['Virtual.txt','Physical.txt','Name.txt']
 
 for x in arrayOfFiles:
-	os.remove(x)
+	if os.path.exists(x):
+		os.remove(x)
 for oneLine in output.split("\n"):
 	if count < 2:
 		count = count+1
